+++ pywwt/jupyter.py
@@ -50,18 +50,8 @@ class WWTJupyterWidget(widgets.DOMWidget, BaseWWTWidget):
     def _send_msg(self, **kwargs):
         self.send(kwargs)
 
-    def load_fits_data(self, filename):
-        """
-        Load a FITS file.
-
-        Parameters
-        ----------
-        filename : str
-            The filename of the FITS file to display.
-        """
-        self._validate_fits_data(filename)
-        url = serve_file(filename, extension='.fits')
-        self._send_msg(event='load_fits', url=url)
+    def _serve_file(self, filename, extension=''):
+        return serve_file(filename, extension=extension)
 
     @property
     def layer_controls(self):