+++ pywwt/tests/test_layers.py
@@ -1,6 +1,9 @@
 import os
 import pytest
 
+import numpy as np
+
+from astropy.wcs import WCS
 from astropy.table import Table
 from astropy import units as u
 from astropy.coordinates import SkyCoord
@@ -24,12 +27,12 @@ class TestLayers:
 
         assert len(self.client.layers) == 0
         assert str(self.client.layers) == 'Layer manager with no layers'
-        layer1 = self.client.layers.add_data_layer(table=self.table)
+        layer1 = self.client.layers.add_table_layer(table=self.table)
         assert len(self.client.layers) == 1
         assert str(self.client.layers) == ('Layer manager with 1 layers:\n\n'
                                            '  [0]: TableLayer with 3 markers\n')
 
-        layer2 = self.client.layers.add_data_layer(table=self.table)
+        layer2 = self.client.layers.add_table_layer(table=self.table)
 
         assert len(self.client.layers) == 2
         assert str(self.client.layers) == ('Layer manager with 2 layers:\n\n'
@@ -56,7 +59,7 @@ class TestLayers:
 
     def test_alt_unit(self):
 
-        layer = self.client.layers.add_data_layer(table=self.table)
+        layer = self.client.layers.add_table_layer(table=self.table)
 
         # Using a string
         layer.alt_unit = 'm'
@@ -87,7 +90,7 @@ class TestLayers:
 
     def test_lon_unit(self):
 
-        layer = self.client.layers.add_data_layer(table=self.table)
+        layer = self.client.layers.add_table_layer(table=self.table)
 
         # Using a string
         layer.lon_unit = 'deg'
@@ -115,7 +118,7 @@ class TestLayers:
 
     def test_alt_type(self):
 
-        layer = self.client.layers.add_data_layer(table=self.table)
+        layer = self.client.layers.add_table_layer(table=self.table)
 
         layer.alt_type = 'depth'
 
@@ -129,7 +132,7 @@ class TestLayers:
         self.table['altitude2'] = [1, 4, 5] * u.def_unit('same_as_km', 1000 * u.m)
         self.table['flux'].unit = u.kg
 
-        layer = self.client.layers.add_data_layer(table=self.table)
+        layer = self.client.layers.add_table_layer(table=self.table)
 
         assert layer.alt_att == ''
         assert layer.alt_unit is None
@@ -153,7 +156,7 @@ class TestLayers:
         self.table['longitude2'] = [1, 4, 5] * u.def_unit('same_as_deg', 3600 * u.arcsec)
         self.table['flux'].unit = u.kg
 
-        layer = self.client.layers.add_data_layer(table=self.table)
+        layer = self.client.layers.add_table_layer(table=self.table)
 
         assert layer.lon_att == 'ra'
         assert layer.lon_unit is u.deg
@@ -174,7 +177,7 @@ class TestLayers:
     def test_update_data(self):
 
         self.table['flux'].unit = 'm'
-        layer = self.client.layers.add_data_layer(table=self.table,
+        layer = self.client.layers.add_table_layer(table=self.table,
                                                   lon_att='ra', lat_att='dec', alt_att='flux')
 
         assert layer.lon_att == 'ra'
@@ -246,7 +249,7 @@ def test_layers_image(tmpdir, wwt_qt_client):
     table['dec'] = [84, 85, 86, 87, 88]
     table['ra'] = [250, 260, 270, 280, 290] * u.deg
 
-    layer1 = wwt.layers.add_data_layer(table=table)
+    layer1 = wwt.layers.add_table_layer(table=table)
 
     # Case where we change the default values on initialization
 
@@ -256,7 +259,7 @@ def test_layers_image(tmpdir, wwt_qt_client):
     table['ra'] = [250, 260, 270, 280, 290] * u.deg
     table['other'] = [255, 265, 275, 285, 295] * u.deg
 
-    layer2 = wwt.layers.add_data_layer(table=table, color='red', lon_att='other', size_scale=100, opacity=0.5)
+    layer2 = wwt.layers.add_table_layer(table=table, color='red', lon_att='other', size_scale=100, opacity=0.5)
 
     # Case where we change the values after initialization
 
@@ -266,7 +269,7 @@ def test_layers_image(tmpdir, wwt_qt_client):
     table['ra'] = [250, 260, 270, 280, 290] * u.deg
     table['other'] = [255, 265, 275, 285, 295] * u.deg
 
-    layer3 = wwt.layers.add_data_layer(table=table)
+    layer3 = wwt.layers.add_table_layer(table=table)
 
     wwt.wait(2)
 
@@ -283,7 +286,7 @@ def test_layers_image(tmpdir, wwt_qt_client):
     table['ra'] = [250, 260, 270, 280, 290] * u.deg
     table['other'] = [255, 265, 275, 285, 295] * u.deg
 
-    layer4 = wwt.layers.add_data_layer(table=table, cmap_att='other', size_att='flux')
+    layer4 = wwt.layers.add_table_layer(table=table, cmap_att='other', size_att='flux')
 
     # Case with size and color encoding where we change the values after initialization
 
@@ -293,7 +296,7 @@ def test_layers_image(tmpdir, wwt_qt_client):
     table['ra'] = [250, 260, 270, 280, 290] * u.deg
     table['other'] = [255, 265, 275, 285, 295] * u.deg
 
-    layer5 = wwt.layers.add_data_layer(table=table)
+    layer5 = wwt.layers.add_table_layer(table=table)
 
     wwt.wait(2)
 
@@ -306,3 +309,77 @@ def test_layers_image(tmpdir, wwt_qt_client):
     # OpenGL features that aren't available there.
     if os.environ.get('CI', 'false').lower() == 'false':
         assert_widget_image(tmpdir, wwt, 'sky_layers.png')
+
+
+def test_image_layer_equ(tmpdir, wwt_qt_client):
+
+    # A series of tests that excercise the image layer functionality and compare
+    # the results with a set of baseline images.
+
+    wwt = wwt_qt_client
+
+    wwt.foreground = 'Black Sky Background'
+    wwt.background = 'Black Sky Background'
+
+    wwt.center_on_coordinates(SkyCoord(30 * u.deg, 40 * u.deg))
+
+    array = np.arange(10000).reshape((100, 100))
+    wcs = WCS()
+    wcs.wcs.ctype = 'RA---TAN', 'DEC--TAN'
+    # wcs.wcs.ctype = 'GLON-CAR', 'GLAT-CAR'
+    wcs.wcs.crval = 30, 40
+    wcs.wcs.crpix = 50.5, 50.5
+    wcs.wcs.cdelt = -0.1, 0.1
+
+    wwt.layers.add_image_layer(image=(array, wcs))
+
+    wwt.wait(2)
+
+    # For now this test doesn't work in CI, seemingly because of some
+    # OpenGL features that aren't available there.
+    if os.environ.get('CI', 'false').lower() == 'false':
+        assert_widget_image(tmpdir, wwt, 'image_layer_equ.png')
+
+
+def test_image_layer_gal(tmpdir, wwt_qt_client):
+
+    # A series of tests that excercise the image layer functionality and compare
+    # the results with a set of baseline images.
+
+    wwt = wwt_qt_client
+
+    wwt.foreground = 'Black Sky Background'
+    wwt.background = 'Black Sky Background'
+
+    array = np.arange(10000).reshape((100, 100))
+    wcs = WCS()
+    wcs.wcs.ctype = 'GLON-CAR', 'GLAT-CAR'
+    wcs.wcs.crpix = 50.5, 50.5
+    wcs.wcs.cdelt = -0.03, 0.03
+
+    wcs.wcs.crval = 33, 43
+    wwt.layers.add_image_layer(image=(array, wcs))
+
+    wcs.wcs.crval = 27, 43
+    layer2 = wwt.layers.add_image_layer(image=(array, wcs))
+    layer2.vmin = -5000
+    layer2.vmax = 15000
+
+    wcs.wcs.crval = 27, 37
+    layer3 = wwt.layers.add_image_layer(image=(array, wcs))
+    layer3.stretch = 'log'
+
+    wcs.wcs.crval = 33, 37
+    layer4 = wwt.layers.add_image_layer(image=(array, wcs))
+    layer4.opacity = 0.5
+
+    wwt.wait(2)
+
+    wwt.center_on_coordinates(SkyCoord(30 * u.deg, 40 * u.deg, frame='galactic'), fov=14 * u.deg)
+
+    wwt.wait(2)
+
+    # For now this test doesn't work in CI, seemingly because of some
+    # OpenGL features that aren't available there.
+    if os.environ.get('CI', 'false').lower() == 'false':
+        assert_widget_image(tmpdir, wwt, 'image_layer_gal.png')